agent_requirements.txt
google-adk>=1.0.0
a2a-sdk>=0.3.0
httpx>=0.27.0
uvicorn>=0.29.0
